+++ src/Resources/contao/config/config.php
@@ -3,4 +3,4 @@
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['getAttributesFromDca']['huhTinyMce'] = [\HeimrichHannot\TinyMceBundle\EventListener\GetAttributesFromDcaListener::class, 'onGetAttributesFromDca'];
+$GLOBALS['TL_HOOKS']['getAttributesFromDca']['huhTinyMce'] = [\HeimrichHannot\TinyMceBundle\EventListener\Contao\GetAttributesFromDcaListener::class, '__invoke'];